move StreamFactory into its own file
